Added build.xml
build.xml is an ANT build script. Apache ANT (Another Neat Tool) is a
utility commonly used for building java projects. If you have ant
installed, you can simply type "ant" in the terminal to run the default
ant task, which in this case is to compile the project. To build a zip
file suitable for submission to the course portal, type "ant dist" and
the file will be created in the dist/ directory if compilation
succeeds.