Introduced a bi-partite matching heuristics.
The heuristics now finds the real distance between all boxes and goals.
Boxes are then matched to goals in such a way to minimize the total path.
This is really slow, but never under-estimates the number of moves
required.

The matching is done by source code I pulled from the Internet.
It is clerly marked with its origin, the code of honor says this is okay
(I think) and it is not really an AI algorithm. Never the less, we should
think about it.